Worked on fully narrowing cell candidates
Just making this the most recent commit message so it's easier to find: Worked on fully narrowing cell candidates by stopping narrowing only when the board is equal to the board with narrowed cell candidates (meaning no more narrowing can be done). Also a practice board has been hard-coded. See my code in main for the problem: When I try to set board2 = board.boardWithNarrowedCandidates, it sets the value of the original board object to board.boardWithNarrowedCandidates as well